third_party/blink/web_tests/external/wpt/editing/run/justifyleft_1001-2000-expected.txt

This is a testharness.js-based test.
Found 62 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div style=text-align:justify><p>[foo<p>bar]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p><p>extra</p>" but got "<div style=\\"text-align:justify\\"><p style=\\"text-align:left\\">foo</p><p style=\\"text-align:left\\">bar</p></div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div style=text-align:justify><p>[foo<p>bar]</div><p>extra" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["justifyleft",""]] "<div align=left><p>[foo]<p>bar</div><p>extra" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "-webkit-left"
[FAIL] [["justifyleft",""]] "<div align=left><p>[foo]<p>bar</div><p>extra" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "-webkit-left"
[FAIL] [["defaultparagraphseparator","div"],["justifyleft",""]] "<div align=left><p>[foo<p>bar}</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p><p>extra</p>" but got "<div align=\\"left\\"><p>foo</p><p>bar</p></div><p>extra</p>"
[FAIL] [["defaultparagraphseparator","div"],["justifyleft",""]] "<div align=left><p>[foo<p>bar}</div><p>extra" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "-webkit-left"
[FAIL] [["defaultparagraphseparator","div"],["justifyleft",""]] "<div align=left><p>[foo<p>bar}</div><p>extra" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "-webkit-left"
[FAIL] [["defaultparagraphseparator","p"],["justifyleft",""]] "<div align=left><p>[foo<p>bar}</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p><p>extra</p>" but got "<div align=\\"left\\"><p>foo</p><p>bar</p></div><p>extra</p>"
[FAIL] [["defaultparagraphseparator","p"],["justifyleft",""]] "<div align=left><p>[foo<p>bar}</div><p>extra" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "-webkit-left"
[FAIL] [["defaultparagraphseparator","p"],["justifyleft",""]] "<div align=left><p>[foo<p>bar}</div><p>extra" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "-webkit-left"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyleft",""]] "<div style=text-align:left><p>[foo<p>bar]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p><p>extra</p>" but got "<div style=\\"text-align:left\\"><p>foo</p><p>bar</p></div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyleft",""]] "<div style=text-align:left><p>[foo<p>bar]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p><p>extra</p>" but got "<div style=\\"text-align:left\\"><p>foo</p><p>bar</p></div><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div style=text-align:left><p>[foo<p>bar]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p><p>extra</p>" but got "<div style=\\"text-align:left\\"><p>foo</p><p>bar</p></div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div style=text-align:left><p>[foo<p>bar]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p><p>extra</p>" but got "<div style=\\"text-align:left\\"><p>foo</p><p>bar</p></div><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyleft",""]] "<div align=right><p>[foo]<p>bar</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div align=\\"right\\"><div style=\\"text-align:left\\"><p>foo</p></div><p>bar</p></div><p>extra</p>" but got "<div align=\\"right\\"><p style=\\"text-align:left\\">foo</p><p>bar</p></div><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyleft",""]] "<div align=right><p>[foo]<p>bar</div><p>extra" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "right" but got "-webkit-right"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyleft",""]] "<div align=right><p>[foo]<p>bar</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div align=\\"right\\"><div style=\\"text-align:left\\"><p>foo</p></div><p>bar</p></div><p>extra</p>" but got "<div align=\\"right\\"><p style=\\"text-align:left\\">foo</p><p>bar</p></div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyleft",""]] "<div align=right><p>[foo]<p>bar</div><p>extra" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "right" but got "-webkit-right"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div align=right><p>[foo]<p>bar</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div align=\\"right\\"><div style=\\"text-align:left\\"><p>foo</p></div><p>bar</p></div><p>extra</p>" but got "<div align=\\"right\\"><p style=\\"text-align:left\\">foo</p><p>bar</p></div><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div align=right><p>[foo]<p>bar</div><p>extra" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "right" but got "-webkit-right"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div align=right><p>[foo]<p>bar</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div align=\\"right\\"><div style=\\"text-align:left\\"><p>foo</p></div><p>bar</p></div><p>extra</p>" but got "<div align=\\"right\\"><p style=\\"text-align:left\\">foo</p><p>bar</p></div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div align=right><p>[foo]<p>bar</div><p>extra" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "right" but got "-webkit-right"
[FAIL] [["defaultparagraphseparator","div"],["justifyleft",""]] "<div align=right><p>[foo<p>bar}</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p><p>extra</p>" but got "<div align=\\"right\\"><p style=\\"text-align:left\\">foo</p><p style=\\"text-align:left\\">bar</p></div><p>extra</p>"
[FAIL] [["defaultparagraphseparator","div"],["justifyleft",""]] "<div align=right><p>[foo<p>bar}</div><p>extra" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "right" but got "-webkit-right"
[FAIL] [["defaultparagraphseparator","p"],["justifyleft",""]] "<div align=right><p>[foo<p>bar}</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p><p>extra</p>" but got "<div align=\\"right\\"><p style=\\"text-align:left\\">foo</p><p style=\\"text-align:left\\">bar</p></div><p>extra</p>"
[FAIL] [["defaultparagraphseparator","p"],["justifyleft",""]] "<div align=right><p>[foo<p>bar}</div><p>extra" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "right" but got "-webkit-right"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyleft",""]] "<div style=text-align:right><p>[foo]<p>bar</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><div style=\\"text-align:left\\"><p>foo</p></div><p>bar</p></div><p>extra</p>" but got "<div style=\\"text-align:right\\"><p style=\\"text-align:left\\">foo</p><p>bar</p></div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyleft",""]] "<div style=text-align:right><p>[foo]<p>bar</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><div style=\\"text-align:left\\"><p>foo</p></div><p>bar</p></div><p>extra</p>" but got "<div style=\\"text-align:right\\"><p style=\\"text-align:left\\">foo</p><p>bar</p></div><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div style=text-align:right><p>[foo]<p>bar</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><div style=\\"text-align:left\\"><p>foo</p></div><p>bar</p></div><p>extra</p>" but got "<div style=\\"text-align:right\\"><p style=\\"text-align:left\\">foo</p><p>bar</p></div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div style=text-align:right><p>[foo]<p>bar</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\"><div style=\\"text-align:left\\"><p>foo</p></div><p>bar</p></div><p>extra</p>" but got "<div style=\\"text-align:right\\"><p style=\\"text-align:left\\">foo</p><p>bar</p></div><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyleft",""]] "<div style=text-align:right><p>[foo<p>bar]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p><p>extra</p>" but got "<div style=\\"text-align:right\\"><p style=\\"text-align:left\\">foo</p><p style=\\"text-align:left\\">bar</p></div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyleft",""]] "<div style=text-align:right><p>[foo<p>bar]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p><p>extra</p>" but got "<div style=\\"text-align:right\\"><p style=\\"text-align:left\\">foo</p><p style=\\"text-align:left\\">bar</p></div><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div style=text-align:right><p>[foo<p>bar]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p><p>extra</p>" but got "<div style=\\"text-align:right\\"><p style=\\"text-align:left\\">foo</p><p style=\\"text-align:left\\">bar</p></div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div style=text-align:right><p>[foo<p>bar]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p><p>extra</p>" but got "<div style=\\"text-align:right\\"><p style=\\"text-align:left\\">foo</p><p style=\\"text-align:left\\">bar</p></div><p>extra</p>"
[FAIL] [["defaultparagraphseparator","div"],["justifyleft",""]] "<div align=left>[foo</div>bar]<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar<p>extra</p>" but got "<div align=\\"left\\">foo</div>bar<p>extra</p>"
[FAIL] [["defaultparagraphseparator","div"],["justifyleft",""]] "<div align=left>[foo</div>bar]<p>extra" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "-webkit-left"
[FAIL] [["defaultparagraphseparator","div"],["justifyleft",""]] "<div align=left>[foo</div>bar]<p>extra" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "-webkit-left"
[FAIL] [["defaultparagraphseparator","p"],["justifyleft",""]] "<div align=left>[foo</div>bar]<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar<p>extra</p>" but got "<div align=\\"left\\">foo</div>bar<p>extra</p>"
[FAIL] [["defaultparagraphseparator","p"],["justifyleft",""]] "<div align=left>[foo</div>bar]<p>extra" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "-webkit-left"
[FAIL] [["defaultparagraphseparator","p"],["justifyleft",""]] "<div align=left>[foo</div>bar]<p>extra" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "-webkit-left"
[FAIL] [["defaultparagraphseparator","div"],["justifyleft",""]] "<div align=left>fo[o</div>b]ar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar<p>extra</p>" but got "<div align=\\"left\\">foo</div>bar<p>extra</p>"
[FAIL] [["defaultparagraphseparator","div"],["justifyleft",""]] "<div align=left>fo[o</div>b]ar<p>extra" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "-webkit-left"
[FAIL] [["defaultparagraphseparator","div"],["justifyleft",""]] "<div align=left>fo[o</div>b]ar<p>extra" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "-webkit-left"
[FAIL] [["defaultparagraphseparator","p"],["justifyleft",""]] "<div align=left>fo[o</div>b]ar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar<p>extra</p>" but got "<div align=\\"left\\">foo</div>bar<p>extra</p>"
[FAIL] [["defaultparagraphseparator","p"],["justifyleft",""]] "<div align=left>fo[o</div>b]ar<p>extra" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "-webkit-left"
[FAIL] [["defaultparagraphseparator","p"],["justifyleft",""]] "<div align=left>fo[o</div>b]ar<p>extra" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "-webkit-left"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyleft",""]] "<div style=text-align:left>[foo</div>bar]<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar<p>extra</p>" but got "<div style=\\"text-align:left\\">foo</div>bar<p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyleft",""]] "<div style=text-align:left>[foo</div>bar]<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar<p>extra</p>" but got "<div style=\\"text-align:left\\">foo</div>bar<p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div style=text-align:left>[foo</div>bar]<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar<p>extra</p>" but got "<div style=\\"text-align:left\\">foo</div>bar<p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div style=text-align:left>[foo</div>bar]<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar<p>extra</p>" but got "<div style=\\"text-align:left\\">foo</div>bar<p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyleft",""]] "<div style=text-align:left>fo[o</div>b]ar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar<p>extra</p>" but got "<div style=\\"text-align:left\\">foo</div>bar<p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyleft",""]] "<div style=text-align:left>fo[o</div>b]ar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar<p>extra</p>" but got "<div style=\\"text-align:left\\">foo</div>bar<p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div style=text-align:left>fo[o</div>b]ar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar<p>extra</p>" but got "<div style=\\"text-align:left\\">foo</div>bar<p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div style=text-align:left>fo[o</div>b]ar<p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar<p>extra</p>" but got "<div style=\\"text-align:left\\">foo</div>bar<p>extra</p>"
[FAIL] [["stylewithcss","true"],["justifyleft",""]] "<span style=text-align:left>[foo]</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<p>extra</p>" but got "<span style=\\"text-align:left\\">foo</span><p>extra</p>"
[FAIL] [["stylewithcss","false"],["justifyleft",""]] "<span style=text-align:left>[foo]</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<p>extra</p>" but got "<span style=\\"text-align:left\\">foo</span><p>extra</p>"
[FAIL] [["stylewithcss","true"],["justifyleft",""]] "<span style=text-align:left>f[o]o</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<p>extra</p>" but got "<span style=\\"text-align:left\\">foo</span><p>extra</p>"
[FAIL] [["stylewithcss","false"],["justifyleft",""]] "<span style=text-align:left>f[o]o</span><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<p>extra</p>" but got "<span style=\\"text-align:left\\">foo</span><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["justifyleft",""]] "<div style=text-align:left>[foo<div style=text-align:left contenteditable=false>bar</div>baz]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<div style=\\"text-align:left\\" contenteditable=\\"false\\">bar</div>baz<p>extra</p>" but got "<div style=\\"text-align:left\\">foo<div style=\\"text-align:left\\" contenteditable=\\"false\\">bar</div>baz</div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["justifyleft",""]] "<div style=text-align:left>[foo<div style=text-align:left contenteditable=false>bar</div>baz]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<div style=\\"text-align:left\\" contenteditable=\\"false\\">bar</div>baz<p>extra</p>" but got "<div style=\\"text-align:left\\">foo<div style=\\"text-align:left\\" contenteditable=\\"false\\">bar</div>baz</div><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div style=text-align:left>[foo<div style=text-align:left contenteditable=false>bar</div>baz]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<div style=\\"text-align:left\\" contenteditable=\\"false\\">bar</div>baz<p>extra</p>" but got "<div style=\\"text-align:left\\">foo<div style=\\"text-align:left\\" contenteditable=\\"false\\">bar</div>baz</div><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["justifyleft",""]] "<div style=text-align:left>[foo<div style=text-align:left contenteditable=false>bar</div>baz]</div><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<div style=\\"text-align:left\\" contenteditable=\\"false\\">bar</div>baz<p>extra</p>" but got "<div style=\\"text-align:left\\">foo<div style=\\"text-align:left\\" contenteditable=\\"false\\">bar</div>baz</div><p>extra</p>"
Harness: the test ran to completion.
